backend/requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Django==4.0
django-celery-results==2.4.0
django-cors-headers==3.13.0
djangorestframework==3.14.0
gunicorn==20.1.0
kombu==5.2.4
prompt-toolkit==3.0.36
python-decouple==3.7
pytz==2022.7.1
redis==4.4.2
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
vine==5.0.0
wcwidth==0.2.6
